backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==4.2.16
django-cors-headers==4.4.0
django-decouple==2.1
djangorestframework==3.15.2
ecdsa==0.19.0
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
